Added file lesson16-1.ts with a new test.
